requirements.txt
langchain
langchain-core

langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy

scikit-learn
streamlit
torch
langchain-groq

langchain-chat-prompt
langchain-prompty
langchain-classic==0.1.0